requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
typing_extensions==4.7.1
yarl==1.9.2
